+++ lib/oli_web/live/insights/insights.ex
@@ -2,6 +2,9 @@ defmodule OliWeb.Insights do
   use Phoenix.LiveView
   alias OliWeb.Insights.{TableHeader, TableRow}
   alias Oli.Authoring.Course
+  alias Oli.Utils
+  alias CSV
+  alias Oli.Activities
 
   def mount(_params, %{"project_slug" => project_slug} = _session, socket) do
     by_activity_rows = Oli.Analytics.ByActivity.query_against_project_slug(project_slug)
@@ -141,4 +144,133 @@ defmodule OliWeb.Insights do
       ""
     end
   end
+
+  def export(project) do
+    filenames =
+      ["raw_analytics.tsv", "by_page.tsv", "by_activity.tsv", "by_objective.tsv"]
+      # CSV Encoder expects charlists for filenames, not strings
+      |> Enum.map(&String.to_charlist(&1))
+
+    analytics =
+      raw_snapshot_data(project.slug)
+      |> Enum.concat(derived_analytics_data(project.slug))
+      |> Enum.map(&CSV.encode(&1, separator: ?\t))
+      |> Enum.map(&Enum.join(&1, ""))
+
+    Enum.zip(filenames, analytics)
+    # Convert to tuples of {filename, CSV table rows}
+    |> Enum.map(&{elem(&1, 0), elem(&1, 1)})
+    |> Utils.zip("analytics.zip")
+  end
+
+  def raw_snapshot_data(project_slug) do
+    snapshots_title_row = [
+      "Activity Title",
+      "Activity Type",
+      "Objective Title",
+      "Attempt Number",
+      "Graded?",
+      "Correct?",
+      "Activity Score",
+      "Activity Out Of",
+      "Hints Requested",
+      "Part Score",
+      "Part Out Of",
+      "Student Response",
+      "Feedback",
+      "Section Title",
+      "Section Slug",
+      "Date Created"
+    ]
+
+    [
+      [
+        snapshots_title_row
+        | Oli.Analytics.Common.snapshots_for_project(project_slug)
+          |> Enum.map(
+            &(&1
+              # Query returns a list of fields
+              # Get activity type
+              |> List.replace_at(1, Activities.get_registration!(Enum.at(&1, 1)).title)
+              # JSON format student response
+              |> List.replace_at(11, Utils.pretty(Enum.at(&1, 11)))
+              # JSON format feedback
+              |> List.replace_at(12, Utils.pretty(Enum.at(&1, 12)))
+              # JSON format date
+              |> List.replace_at(15, Utils.format_datetime(Enum.at(&1, 15))))
+          )
+      ]
+    ]
+  end
+
+  def derived_analytics_data(project_slug) do
+    analytics_title_row = [
+      "Resource Title",
+      "Activity Title",
+      "Number of Attempts",
+      "Relative Difficulty",
+      "Eventually Correct",
+      "First Try Correct"
+    ]
+
+    [
+      Oli.Analytics.ByPage.query_against_project_slug(project_slug),
+      Oli.Analytics.ByActivity.query_against_project_slug(project_slug),
+      Oli.Analytics.ByObjective.query_against_project_slug(project_slug)
+    ]
+    |> Enum.map(&[analytics_title_row | extract_analytics(&1)])
+  end
+
+  def extract_analytics([
+        %{
+          slice: slice,
+          number_of_attempts: number_of_attempts,
+          relative_difficulty: relative_difficulty,
+          eventually_correct: eventually_correct,
+          first_try_correct: first_try_correct
+        } = h
+        | t
+      ]) do
+    [
+      [
+        slice.title,
+        if !Map.has_key?(h, :activity) do
+          slice.title
+        else
+          h.activity.title
+        end,
+        if is_nil(number_of_attempts) do
+          "No attempts"
+        else
+          Integer.to_string(number_of_attempts)
+        end,
+        if is_nil(relative_difficulty) do
+          ""
+        else
+          Float.to_string(truncate(relative_difficulty))
+        end,
+        if is_nil(eventually_correct) do
+          ""
+        else
+          format_percent(eventually_correct)
+        end,
+        if is_nil(first_try_correct) do
+          ""
+        else
+          format_percent(first_try_correct)
+        end
+      ]
+      | extract_analytics(t)
+    ]
+  end
+
+  def extract_analytics([]), do: []
+
+  def truncate(float_or_nil) when is_nil(float_or_nil), do: nil
+  def truncate(float_or_nil) when is_float(float_or_nil), do: Float.round(float_or_nil, 2)
+
+  def format_percent(float_or_nil) when is_nil(float_or_nil), do: nil
+
+  def format_percent(float_or_nil) when is_float(float_or_nil),
+    do: "#{round(100 * float_or_nil)}%"
 end